LPS-44959 Rename
