+++ modules/portal/portal-sso-facebook-connect/bnd.bnd
@@ -1,0 +1,7 @@
+Bundle-Name: Liferay Portal SSO Facebook Connect
+Bundle-SymbolicName: com.liferay.portal.sso.facebook.connect
+Bundle-Version: 1.0.0
+Include-Resource:\
+	classes,\
+	META-INF/resources=src/META-INF/resources
+Web-ContextPath: /portal-sso-facebook-connect
