+++ modules/portal/portal-sso-facebook-connect/build.xml
