+++ modules/portal/portal-sso-facebook-connect/ivy.xml
