+++ modules/portal/portal-sso-facebook-connect/src/META-INF/resources/html/portlet/login/navigation/facebook.jsp
