--- modules/portal/portal-sso-facebook/bnd.bnd
@@ -1,7 +1,0 @@
-Bundle-Name: Liferay Portal SSO Facebook
-Bundle-SymbolicName: com.liferay.portal.sso.facebook
-Bundle-Version: 1.0.0
-Include-Resource:\
-	classes,\
-	META-INF/resources=src/META-INF/resources
-Web-ContextPath: /portal-sso-facebook
